+++ backend_new/src/main/java/com/app/persistence/repository/UserRepository.java
@@ -19,10 +19,10 @@ public interface UserRepository extends ReactiveCrudRepository<UserEntity, Long>
     Mono<UserEntity> findByRecoveryToken(String token);
 
     @Query("SELECT COUNT(*) FROM users WHERE username = :username")
-    Mono<Long> existsByUsername(String username);
+    Mono<Integer> existsByUsername(String username);
 
     @Query("SELECT COUNT(*) FROM users WHERE email = :email")
-    Mono<Long> existsByEmail(String email);
+    Mono<Integer> existsByEmail(String email);
 
     @org.springframework.data.r2dbc.repository.Modifying
     @Query("UPDATE users SET is_enabled = true, verification_token = NULL WHERE verification_token = :token")
